+++ app/src/main/java/code/with/me/testroomandnavigationdrawertest/domain/repo/FolderRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface FolderRepository {
     fun getAllFolders(): Flow<List<Folder>>
-    fun getNotesInFolder(folderId: Int): Flow<List<Note>>
+    fun getNotesInFolder(folderId: Long): Flow<List<Note>>
     fun getAllFoldersSortByNameASC(): Flow<List<Folder>>
     fun getAllFoldersSortByNameDESC(): Flow<List<Folder>>
     fun getAllFoldersLastOpenedNewest(): Flow<List<Folder>>
@@ -14,7 +14,7 @@ interface FolderRepository {
     fun getAllFoldersLastEditedNewest(): Flow<List<Folder>>
     fun getAllFoldersLastEditedOldest(): Flow<List<Folder>>
 
-    fun updateLastOpenedFolder(time: Long, folderId: Int)
+    fun updateLastOpenedFolder(time: Long, folderId: Long)
     suspend fun getAllFoldersFavorite(): Flow<List<Folder>>
     suspend fun insertFolder(folder: Folder): Long
     suspend fun updateFolder(folder: Folder)
