+++ app/src/main/java/code/with/me/testroomandnavigationdrawertest/ui/base/BaseFolderListFragment.kt
@@ -142,7 +142,7 @@ abstract class BaseFolderListFragment :
         }
     }
 
-    private fun openDetailFragment(id: Int) {
+    private fun openDetailFragment(id: Long) {
         lifecycleScope.launch {
             withContext(Dispatchers.IO) {
                 async {
